requirements.txt
neo4j
qdrant-client
sentence-transformers
google-generativeai
fastapi
uvicorn
pandas
pypdf
python-dotenv
pytest
httpx
